Add files via upload
The given I-Phone csv file includes the collected data from the flipkart though the process of web scraping. Through this collection of data we can analyze the data and make a decision that which mobile model has the highest buying rate so that we can suggest our client to list those mobile on their platform.